Refactor coordinate systems to include a generic Kind parameter in Cartesian and Spherical Coords (mod required to avoid colision)
