+++ src/coordinates/mod.rs
@@ -62,6 +62,7 @@ pub mod cartesian;
 pub mod spherical;
 pub mod frames;
 pub mod centers;
+pub mod kinds;
 
-pub use cartesian::*;
-pub use spherical::*;
+//pub use cartesian::*;
+//pub use spherical::*;
